totp cli #27
added in commands to interface with totp data on a server. not much is
needed for this but other checks could be put in place so that data is
not created when there is no reason for it to be like preventing
recovery keys from being created if no totp is enabled and maybe also
preven totp from being enabled if no password is found.

+++ rfs-server/src/routing/handle/auth/totp/mod.rs
@@ -104,7 +104,7 @@ pub async fn post(
                 &pg_algo,
                 &secret.as_slice(),
                 &(digits as i32),
-                &(step as i64)
+                &(step as i32)
             ]
         ).await?;
     }
